+++ index.html
@@ -236,5 +236,6 @@
     </template>
 
 </body>
-<script src="js/main.js"></script>
+<script src="./js/main.js" type="module"></script>
+<script src="./js/functions.js"></script>
 </html>
